+++ PARD/Sources/User/ViewControllers/MainLoginViewController.swift
@@ -40,7 +40,6 @@ class MainLoginViewController: UIViewController {
         view.addSubview($0)
     }
     
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         setUpUI()
@@ -107,24 +106,5 @@ class MainLoginViewController: UIViewController {
         controller.performRequests()
         controller.delegate = self
         controller.presentationContextProvider = self
-    }
-    
-    @objc func handleGoogleLogin() {
-        let signInConfig = GIDConfiguration.init(clientID: "215579567587-3qckigpku02urbubjtq4qd9oonpltkt3.apps.googleusercontent.com")
-        GIDSignIn.sharedInstance.signIn(with: signInConfig, presenting: self) { user, error in
-            guard error == nil else { return }
-            guard let user = user else { return }
-            
-            let emailAddress = user.profile?.email
-            print("email Address = \(String(describing: emailAddress))")
-            let fullName = user.profile?.name
-            print("fullName = \(String(describing: fullName))")
-            let givenName = user.profile?.givenName
-            print("givenName = \(String(describing: givenName))")
-            let familyName = user.profile?.familyName
-            print("familyName = \(String(describing: familyName))")
-            let profilePicUrl = user.profile?.imageURL(withDimension: 320)
-            print("profilePicUrl = \(String(describing: profilePicUrl))")
-        }
     }
 }
